Move Exception to Exceptions, cleanup exception redundancies
