--- src/Exception/Api/NotFoundException.php
@@ -1,16 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This software may be modified and distributed under the terms
- * of the MIT license. See the LICENSE file for details.
- */
-
-namespace CarAndClassic\TalkJS\Exception\Api;
-
-use CarAndClassic\TalkJS\Exception\ApiException;
-
-final class NotFoundException extends \Exception implements ApiException
-{
-}
