--- src/Exception/Api/UnknownErrorException.php
@@ -1,41 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This software may be modified and distributed under the terms
- * of the MIT license. See the LICENSE file for details.
- */
-
-namespace CarAndClassic\TalkJS\Exception\Api;
-
-use CarAndClassic\TalkJS\Exception\ApiException;
-use Symfony\Contracts\HttpClient\ResponseInterface;
-
-final class UnknownErrorException extends \Exception implements ApiException
-{
-    protected $response;
-
-    public function __construct(ResponseInterface $response)
-    {
-        $this->response = $response;
-        $content = json_decode($response->getContent(false), true);
-
-        if (!isset($content['reasons'])) {
-            parent::__construct('Unknown error: No reason.');
-
-            return;
-        }
-
-        $field = array_key_first($content['reasons']);
-        $reasons = reset($content['reasons']);
-        $reason = \is_array($reasons) ? $reasons[0] : $reasons;
-
-        parent::__construct("Unknown error: Field $field: $reason.");
-    }
-
-    public function getResponse(): ResponseInterface
-    {
-        return $this->response;
-    }
-}
