+++ src/Exceptions/Api/NotFoundException.php
@@ -7,8 +7,8 @@ declare(strict_types=1);
  * of the MIT license. See the LICENSE file for details.
  */
 
-namespace CarAndClassic\TalkJS;
+namespace CarAndClassic\TalkJS\Exceptions\Api;
 
-interface Exception extends \Throwable
+final class NotFoundException extends \Exception
 {
 }
