+++ src/Exceptions/Api/UnauthorizedException.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This software may be modified and distributed under the terms
+ * of the MIT license. See the LICENSE file for details.
+ */
+
+namespace CarAndClassic\TalkJS\Exceptions\Api;
+
+final class UnauthorizedException extends \Exception
+{
+}
